requirements.txt
# EPS Account Intelligence Agent Dependencies

# Databricks & MLflow
databricks-sdk>=0.20.0
databricks-agents>=0.1.0
mlflow>=3.1.0

# LLM Client
databricks-openai>=0.1.0
openai>=1.0.0

# HTTP Client
httpx>=0.25.0

# Retry Logic
backoff>=2.2.0

# Type Hints
pydantic>=2.0.0

